pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Backend for Demo CSA Management App"
authors = ["Brian Ibbotson <brian@quinceleaf.dev>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2"
psycopg2-binary = "^2.8.6"
gunicorn = "^20.1.0"
django-redis = "^4.12.1"
django-filter = "^2.4.0"
django-import-export = "^2.5.0"
django-simple-history = "^2.12.0"
djangorestframework = "^3.12.2"
Pillow = "^8.1.2"
python-dotenv = "^0.15.0"
python-ulid = "^1.0.2"
pytz = "^2021.1"
celery = "5.0.5"
coverage = "^5.5"
django-celery-beat = "^2.2.0"
django-coverage-plugin = "^1.8.0"
weasyprint = "^52.5"
djangorestframework-simplejwt = "^4.7.0"
django-cors-headers = "^3.7.0"
PyJWT = "^2.1.0"

[tool.poetry.dev-dependencies]
black = "^21.7b0"
flake8 = "^3.9.2"
model-bakery = "^1.2.1"
selenium = "^3.141.0"
PyNaCl = "^1.4.0"
Faker = "^8.10.1"
requests = "^2.26.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
